==> src/1.load_packages.R <==
# Load Packges
library(dplyr)
library(pls)
library(caret)
library(ggplot2)
library(gridExtra)
library(vroom)
#devtools::install_github(repo = "plantphys/spectratrait", dependencies=TRUE)
library(spectratrait)
library(ggdist)
library(ggthemes)
library(gghalves)
